requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-crontab==0.7.1
django-imagekit==4.1.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.1
Faker==15.0.0
frozenlist==1.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
multidict==6.0.2
mymelipayamak==1.0.0
packaging==21.3
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toposort==1.7
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
zeep==4.2.0
zope.interface==5.5.0
aiosignal==1.2.0
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-crontab==0.7.1
django-imagekit==4.1.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.1
Faker==15.0.0
frozenlist==1.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
multidict==6.0.2
mymelipayamak==1.0.0
packaging==21.3
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toposort==1.7
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
zeep==4.2.0
zope.interface==5.5.0
